refactor : 팩토리메서드에서만 도메인의 생성자를 쓸 수 있게 개선
before : 도메인 객체의 생성자가 public으로 열려있었음
after : 생성자를 protected로 바꾸고, 팩토리클래스를 모두 도메인 객체와 같은 패키지로 이동함
테스트 통과
