+++ src/main/java/racingcar/Application.java
@@ -1,7 +1,7 @@
 package racingcar;
 
 import racingcar.controller.RacingCarController;
-import racingcar.factory.ControllerFactory;
+import racingcar.controller.ControllerFactory;
 
 public class Application {
     public static void main(String[] args) {
