+++ src/main/java/racingcar/controller/ControllerFactory.java
@@ -1,6 +1,5 @@
-package racingcar.factory;
+package racingcar.controller;
 
-import racingcar.controller.RacingCarController;
 import racingcar.view.CarsOutputView;
 
 public class ControllerFactory {
