+++ src/main/java/racingcar/domain/Cars.java
@@ -10,7 +10,7 @@ public class Cars {
     private final List<Car> cars;
     private List<Observer> observers;
 
-    public Cars(List<Car> cars) {
+    protected Cars(List<Car> cars) {
         this.cars = cars;
         this.observers = new ArrayList<>();
     }
