+++ src/main/java/racingcar/domain/CarsFactory.java
@@ -1,14 +1,12 @@
-package racingcar.factory;
+package racingcar.domain;
 
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
-import racingcar.domain.Car;
-import racingcar.domain.Cars;
 import racingcar.validation.CarsValidator;
 import racingcar.validation.Validator;
 
-public class CarsFactory {
+public class CarsFactory{
     private static final String CAR_NAMES_DELIMITER = ",";
     private static final int SPLIT_NO_LIMIT = -1;
     private static final Validator<List<Car>> validator = new CarsValidator();
@@ -24,5 +22,4 @@ public class CarsFactory {
                 .map(CarFactory::createCar)
                 .collect(Collectors.toList());
     }
-
 }
