+++ src/test/java/racingcar/domain/CarsTest.java
@@ -7,7 +7,6 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import org.junit.jupiter.api.Test;
-import racingcar.factory.CarsFactory;
 
 public class CarsTest {
     @Test
