+++ src/test/java/racingcar/domain/MoveNumberTest.java
@@ -6,7 +6,6 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import racingcar.factory.MoveNumberFactory;
 
 public class MoveNumberTest {
     @Test
